requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2>=2.8.6,<=2.9.9
drf-spectacular>=0.24.2,<0.25.0
Pillow>=9.3.0,<9.5.0
